+++ tests/resources/tileset.png
